// frontend/web/src/services/API/proserve-wb-packaging-api/models/ComponentVersionTestExecutionSummary.ts
export interface ComponentVersionTestExecutionSummary {
    /**
     * Component version ID
     * @type {string}
     * @memberof ComponentVersionTestExecutionSummary
     */
    componentVersionId: string;
    /**
     * Test instance operating system architecture
     * @type {string}
     * @memberof ComponentVersionTestExecutionSummary
     */
    instanceArchitecture: string;
    /**
     * Test instance ID
     * @type {string}
     * @memberof ComponentVersionTestExecutionSummary
     */
    instanceId: string;
    /**
     * Test instance image upstream (AMI) ID
     * @type {string}
     * @memberof ComponentVersionTestExecutionSummary
     */
    instanceImageUpstreamId: string;
    /**
     * Test instance operating system version
     * @type {string}
     * @memberof ComponentVersionTestExecutionSummary
     */
    instanceOsVersion: string;
    /**
     * Test instance platform
     * @type {string}
     * @memberof ComponentVersionTestExecutionSummary
     */
    instancePlatform: string;
    /**
     * Status of the test execution (PENDING/RUNNING/SUCCESS/FAILED)
     * @type {string}
     * @memberof ComponentVersionTestExecutionSummary
     */
    status: string;
    /**
     * Test execution ID
     * @type {string}
     * @memberof ComponentVersionTestExecutionSummary
     */
    testExecutionId: string;
    /**
     * Create Date
     * @type {string}
     * @memberof ComponentVersionTestExecutionSummary
     */
    createDate: string;
    /**
     * Last Update Date
     * @type {string}
     * @memberof ComponentVersionTestExecutionSummary
     */
    lastUpdateDate: string;
}